# data/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Athena CTF",
    "description": "# Oxide\n\nOxide is the api server of [Athena CTF](https://github.com/athena-ctf/athena-ctf) platform.\n\n## Usage\n\n```\noxide run /data/config.toml\n```\n",
    "contact": {
      "name": "AthenaCTF Team",
      "email": "athena.help@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://www.mit.edu/~amini/LICENSE.md"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/achivement/{id}": {
      "get": {
        "tags": [
          "achievement"
        ],
        "summary": "Retrieve achievement by id",
        "operationId": "retrieve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved achievement by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No achievement found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/flag/verify": {
      "post": {
        "tags": [
          "flag"
        ],
        "summary": "Verify flag",
        "operationId": "verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyFlagSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Verified flag successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagVerificationResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/instance/{id}": {
      "get": {
        "tags": [
          "instance"
        ],
        "summary": "Retrieve instance by id",
        "operationId": "retrieve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved instance by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No instance found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "instance"
        ],
        "summary": "Delete instance by id",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted instance by id successfully"
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No instance found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "instance"
        ],
        "summary": "Update instance by id",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated instance by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters/request body format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No instance found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/invite/{id}": {
      "get": {
        "tags": [
          "invite"
        ],
        "summary": "Retrieve invite by id",
        "operationId": "retrieve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved invite by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No invite found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "invite"
        ],
        "summary": "Delete invite by id",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted invite by id successfully"
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No invite found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "invite"
        ],
        "summary": "Update invite by id",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated invite by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters/request body format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No invite found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AchievementModel": {
        "type": "object",
        "required": [
          "id",
          "value",
          "date_created",
          "prize",
          "challenge_id"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "date_created": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "player_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "prize": {
            "type": "integer",
            "format": "int32"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "ErrorModel": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "FlagVerificationResult": {
        "type": "object",
        "required": [
          "result"
        ],
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ResultEnum"
          }
        }
      },
...
      "InstanceDetails": {
        "type": "object",
        "required": [
          "challenge_id",
          "player_id"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "player_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "InstanceModel": {
        "type": "object",
        "required": [
          "id",
          "date_created",
          "container_id",
          "container_port",
          "expiry",
          "challenge_id",
          "player_id"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "container_id": {
            "type": "string"
          },
          "container_port": {
            "type": "integer",
            "format": "int64"
          },
          "date_created": {
            "type": "string",
            "format": "date-time"
          },
          "expiry": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "player_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "InviteDetails": {
        "type": "object",
        "required": [
          "remaining",
          "team_id",
          "usages"
        ],
        "properties": {
          "remaining": {
            "type": "integer",
            "format": "int32"
          },
          "team_id": {
            "type": "string",
            "format": "uuid"
          },
          "usages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "InviteModel": {
        "type": "object",
        "required": [
          "id",
          "usages",
          "remaining",
          "team_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "remaining": {
            "type": "integer",
            "format": "int32"
          },
          "team_id": {
            "type": "string",
            "format": "uuid"
          },
          "usages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ResultEnum": {
        "type": "string",
        "enum": [
          "fail",
          "pass"
        ]
      },
...
      "VerifyFlagSchema": {
        "type": "object",
        "required": [
          "challenge_id",
          "flag"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "flag": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}